PthonResume/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-common==1.1.28
azure-functions==1.12.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.11.1
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==39.0.0
cymem==2.0.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
geographiclib==2.0
geopy==2.3.0
idna==3.4
indian-address-parser @ https://github.com/shriekdj/indian_address_parser/releases/download/v0.1.0/indian_address_parser-0.1.2-py3-none-any.whl
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdfminer.six==20220524
preshed==3.0.8
pyap==0.3.1
pycparser==2.21
pydantic==1.10.4
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==0.10.1


